Start refactor of epoch.rs
