+++ src/lib.rs
@@ -110,6 +110,3 @@ pub mod python;
 
 #[cfg(feature = "std")]
 extern crate core;
-
-#[cfg(feature = "std")]
-mod system_time;
